+++ test/fixtures/privileges.sql
@@ -53,6 +53,9 @@ GRANT ALL ON TABLE
     , images_base64
     , w_or_wo_comma_names
     , items_with_different_col_types
+    , entities
+    , child_entities
+    , grandchild_entities
 TO postgrest_test_anonymous;
 
 GRANT INSERT ON TABLE insertonly TO postgrest_test_anonymous;